abstract customValidator function
