+++ server/app.js
@@ -2,10 +2,9 @@ import express from 'express';
 import bodyParser from 'body-parser';
 import expressValidator from 'express-validator';
 import swaggerUi from 'swagger-ui-express';
-import path from 'path';
 import dotenv from 'dotenv';
 import routes from './routes/index';
-
+import customValidator from './middlewares/customValidator';
 
 const swaggerDocument = require('../swagger.json');
 
@@ -17,22 +16,7 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
 // Validator to check requests
-app.use(expressValidator({
-  customValidators: {
-    isImage(imageName) {
-      const imageExtension = (path.extname(imageName)).toLowerCase();
-      switch (imageExtension) {
-        case '.jpg':
-          return '.jpg';
-        case '.jpeg':
-          return '.jpeg';
-        case '.png':
-          return '.png';
-        default:
-          return false;
-      }
-    },
-  } }));
+app.use(expressValidator(customValidator));
 
 // Versioning and Routes
 app.use('/api/v1/', routes);
